requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3

django-filter~=22.1
APScheduler~=3.9.1.post1
requests~=2.28.1
celery~=5.2.7
django-celery-beat
django-celery-results
python-decouple
redis